+++ lib/screens/charts/pie/edit/edit_info.dart
@@ -2,6 +2,7 @@ import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
 
 import '../../../../models/chart.dart';
+import '../../../../langs/lang.dart';
 import '../../../../widgets/tiles.dart';
 import '../../../../widgets/danger_zone.dart';
 import '../../../../utils/theme.dart';
@@ -36,6 +37,7 @@ class _EditInfoState extends State<EditInfo>
   @override
   Widget build(BuildContext context) {
     super.build(context);
+    BaseLocalization loc = Localization.currentLocalization;
     return ListView(
       padding: EdgeInsets.symmetric(vertical: 10) + EdgeInsets.only(bottom: 60),
       children: [
@@ -43,11 +45,11 @@ class _EditInfoState extends State<EditInfo>
           controller: widget.nameController,
           decoration: InputDecoration(
             contentPadding: EdgeInsets.symmetric(horizontal: 10),
-            labelText: 'Name',
+            labelText: loc.name,
           ),
           autovalidateMode: AutovalidateMode.always,
           validator: (text) {
-            if (text == null || text.isEmpty) return 'The name can NOT be null';
+            if (text == null || text.isEmpty) return loc.canNotBeEmpty;
             return null;
           },
           onFieldSubmitted: (text) {
@@ -59,27 +61,7 @@ class _EditInfoState extends State<EditInfo>
         SliderListTile(
           subtitle: Row(
             children: [
-              Text('Center Space Radius'),
-              Spacer(),
-              Text(
-                widget.chart.data.centerSpaceRadius.toStringAsPrecision(3) +
-                    '/100',
-              ),
-            ],
-          ),
-          value: widget.chart.data.centerSpaceRadius,
-          max: 100,
-          count: SizedBox(),
-          onChanged: (v) {
-            widget.chart.data =
-                widget.chart.data.copyWith(centerSpaceRadius: v);
-            widget.requestUpdate(widget.chart);
-          },
-        ),
-        SliderListTile(
-          subtitle: Row(
-            children: [
-              Text('Rotation degree'),
+              Text(loc.rotationDegree),
               Spacer(),
               Text(
                 widget.chart.data.startDegreeOffset.toStringAsPrecision(3) +
@@ -99,7 +81,7 @@ class _EditInfoState extends State<EditInfo>
         SliderListTile(
           subtitle: Row(
             children: [
-              Text('Sections Space'),
+              Text(loc.sectionsSpace),
               Spacer(),
               Text(
                 widget.chart.data.sectionsSpace.toStringAsPrecision(3) + '/100',
@@ -115,7 +97,36 @@ class _EditInfoState extends State<EditInfo>
           },
         ),
         ColorPickerListTile(
-          title: 'Center space color',
+          title: loc.backgroundColor,
+          color: widget.chart.backgroundColor,
+          onChange: (color) {
+            widget.chart.backgroundColor = color;
+            widget.requestUpdate(widget.chart);
+          },
+        ),
+        _buildTitle(context, Text(loc.center)),
+        SliderListTile(
+          subtitle: Row(
+            children: [
+              Text(loc.centerSpaceRadius),
+              Spacer(),
+              Text(
+                widget.chart.data.centerSpaceRadius.toStringAsPrecision(3) +
+                    '/100',
+              ),
+            ],
+          ),
+          value: widget.chart.data.centerSpaceRadius,
+          max: 100,
+          count: SizedBox(),
+          onChanged: (v) {
+            widget.chart.data =
+                widget.chart.data.copyWith(centerSpaceRadius: v);
+            widget.requestUpdate(widget.chart);
+          },
+        ),
+        ColorPickerListTile(
+          title: loc.centerSpaceColor,
           color: widget.chart.data.centerSpaceColor,
           onChange: (color) {
             widget.chart.data =
@@ -123,17 +134,9 @@ class _EditInfoState extends State<EditInfo>
             widget.requestUpdate(widget.chart);
           },
         ),
-        ColorPickerListTile(
-          title: 'Background color',
-          color: widget.chart.backgroundColor,
-          onChange: (color) {
-            widget.chart.backgroundColor = color;
-            widget.requestUpdate(widget.chart);
-          },
-        ),
         _buildTitle(
           context,
-          Text('Border'),
+          Text(loc.border),
           trailing: Checkbox(
             value: widget.chart.data.borderData.show,
             onChanged: (value) => updateBorder(show: value),
@@ -142,7 +145,7 @@ class _EditInfoState extends State<EditInfo>
         SliderListTile(
           subtitle: Row(
             children: [
-              Text('Value'),
+              Text(loc.value),
               Spacer(),
               Text(
                 widget.chart.data.borderData.border.top.width
@@ -157,15 +160,15 @@ class _EditInfoState extends State<EditInfo>
           onChanged: (v) => updateBorder(width: v),
         ),
         ColorPickerListTile(
-          title: 'Border color',
+          title: loc.borderColor,
           color: widget.chart.data.borderData.border.top.color,
           onChange: (color) => updateBorder(color: color),
         ),
         _buildTitle(
           context,
-          Text('Sections'),
+          Text(loc.sections),
           trailing: IconButton(
-            tooltip: 'Add',
+            tooltip: loc.add,
             icon: Icon(Icons.add),
             onPressed: () {
               widget.chart.data.sections.add(PieChartSectionData(
@@ -178,14 +181,7 @@ class _EditInfoState extends State<EditInfo>
             splashRadius: 16,
           ),
         ),
-        if (widget.chart.data.sections.isEmpty)
-          Padding(
-            padding: EdgeInsets.all(10),
-            child: Text(
-              'You have no data yet.\nPress on + to create a new one',
-            ),
-          )
-        else ...[
+        if (widget.chart.data.sections.isNotEmpty) ...[
           ...List.generate(
             widget.chart.data.sections.length,
             (index) {
@@ -211,7 +207,7 @@ class _EditInfoState extends State<EditInfo>
             },
           ),
           Text(
-            'Slide to any sides to delete a section',
+            loc.slideToSideToDeleteSection,
             textAlign: TextAlign.center,
             style: Theme.of(context).textTheme.caption,
           ),