+++ lib/widgets/leave.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 
+import '../langs/lang.dart';
+
 class SureWantToLeave extends StatelessWidget {
   const SureWantToLeave({
     Key key,
@@ -12,25 +14,24 @@ class SureWantToLeave extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    BaseLocalization loc = Localization.currentLocalization;
     return AlertDialog(
-      title: Text('Are you sure?'),
-      content: Text(
-        'Are you sure you want to leave without saving your progress?',
-      ),
+      title: Text(loc.areYouSure),
+      content: Text(loc.areYouSureDescription),
       contentPadding: EdgeInsets.symmetric(horizontal: 24, vertical: 5),
       actions: [
         FlatButton(
-          child: Text('Leave'),
+          child: Text(loc.leave),
           color: Colors.redAccent,
           onPressed: onLeave,
         ),
         FlatButton(
-          child: Text('Save and Leave', style: TextStyle(color: Colors.white)),
+          child: Text(loc.saveAndLeave, style: TextStyle(color: Colors.white)),
           color: Colors.teal,
           onPressed: onSaveAndLeave,
         ),
         FlatButton(
-          child: Text('Cancel', style: TextStyle(color: Colors.white)),
+          child: Text(loc.dismiss, style: TextStyle(color: Colors.white)),
           color: Colors.blue,
           onPressed: () => Navigator.pop(context),
         ),